refactor(client): extract version tests to sibling file
